@@ -1,7 +1,7 @@
 {
     "name": "barbieswimcrew/zip-code-validator",
     "description": "Constraint class for international zipcode validation",
-    "keywords": ["symfony", "symfony4", "symfony3", "symfony2", "form", "constraints", "constraint", "validator", "zip code", "zipcode", "postal code", "postalcode", "validation"],
+    "keywords": ["symfony", "form", "constraints", "constraint", "validator", "zip code", "zipcode", "postal code", "postalcode", "validation"],
     "license": "MIT",
     "authors": [
         {
@@ -11,16 +11,20 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=5.6.0",
-        "symfony/validator": ">=2.1"
+        "php": ">=8.0",
+        "symfony/validator": ">=4.4.40"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
-            "ZipCodeValidator\\": "src/ZipCodeValidator",
-            "Tests\\Fixtures\\": "tests/Fixtures"
+            "ZipCodeValidator\\": "src/ZipCodeValidator"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "ZipCodeValidator\\Tests\\": "tests/"
         }
     }
 }
